fix controller and view
