--- solutions/devsprint-caio-santos-3/FinanceApp/Modules/ContactList/ContactListView.swift
@@ -1,55 +1,0 @@
-//
-//  ContactListView.swift
-//  FinanceApp
-//
-//  Created by Rodrigo Borges on 30/12/21.
-//
-
-import Foundation
-import UIKit
-
-class ContactListView: UIView {
-    
-    lazy var tableView: UITableView = {
-
-        let tableView = UITableView(frame: .zero)
-        tableView.translatesAutoresizingMaskIntoConstraints = false
-        tableView.register(ContactCellView.self, forCellReuseIdentifier: ContactCellView.cellIdentifier)
-        return tableView
-    }()
-
-    init(tableViewDataSource: UITableViewDataSource, tableViewDelegate: UITableViewDelegate) {
-        super.init(frame: .zero)
-        tableView.dataSource = tableViewDataSource
-        tableView.delegate = tableViewDelegate
-        backgroundColor = .white
-        addSubviews()
-        configureConstraints()
-
-        tableView.reloadData()
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-extension ContactListView {
-
-    func addSubviews() {
-
-        addSubview(tableView)
-    }
-
-    func configureConstraints() {
-
-        NSLayoutConstraint.activate([
-
-            tableView.leadingAnchor.constraint(equalTo: safeAreaLayoutGuide.leadingAnchor),
-            tableView.trailingAnchor.constraint(equalTo: safeAreaLayoutGuide.trailingAnchor),
-            tableView.topAnchor.constraint(equalTo: safeAreaLayoutGuide.topAnchor),
-            tableView.bottomAnchor.constraint(equalTo: safeAreaLayoutGuide.bottomAnchor),
-        ])
-    }
-}
-
